Add Sentinel-1 EW. Renamed tests for clarity
